korrekte preis anzeige

+++ src/servlets/ArtikelLadenMen.java
@@ -56,7 +56,7 @@ public class ArtikelLadenMen extends HttpServlet {
 		dispatcher.forward(request, response);
 	}
 
-	// Fast eins zu eins aus dem JDBC-Script, S. 19 übernommen
+	// Fast eins zu eins aus dem JDBC-Script, S. 19 ï¿½bernommen
 	private List<ArtikelBean> loadArtikel(Integer art_kategorie) throws ServletException {
 
 		List<ArtikelBean> artikel = new ArrayList<ArtikelBean>();
@@ -77,7 +77,7 @@ public class ArtikelLadenMen extends HttpServlet {
 					String artikelbez = rs.getString("artikelbezeichnung");
 					artikelbean.setArtikelbezeichnung(artikelbez);
 
-					Double preis = Double.valueOf(rs.getLong("preis"));
+					Double preis = Double.valueOf(rs.getDouble("preis"));
 					artikelbean.setPreis(preis);
 
 					Integer kategorie = rs.getInt("kategorie");
